Refactor doxygen parser into subpackage
